+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Guide: https://keepachangelog.com/en/1.0.0/
 -->
 
+## [Unreleased]
+
+- [Core] Experimental. Add support for searching parking locations with access to dynamic and static parking metadata.
+
 ## 2.14.0
 
 - [Core] Update dependencies.